Moved context to root of src/client updated import
