+++ src/client/App.js
@@ -3,7 +3,7 @@ import { Route, Routes, useNavigate, BrowserRouter } from "react-router-dom";
 // import Login from "./components/loginwindow.js";
 // import Signup from "./components/signupwindow.js";
 import Chatboard from './components/Chatboard.jsx';
-import { SocketContext } from './components/Context.js';
+import { SocketContext } from './Context.js';
 import { io } from "socket.io-client";
 
 const socket = io("ws://localhost:3001");
