+++ src/client/components/Chatcategory.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Chatroom from "./Chatroom";
 import { useEffect, useContext } from "react";
-import { SocketContext } from "./Context";
+import { SocketContext } from "../Context";
 
 function Chatcategory(props) {
   const { socket } = useContext(SocketContext);
